Personal Math Class
This project is a Java program I made that replicates a few methods from the Math class. It does not contain all the methods found in the Math class, only a few I found to be the most helpful to me such as pow, sin, cos, toRadian. I also implemented a factorial method that solves "n!" problems and a capital to lowercase converter method.